backend/requirements.txt
fastapi>=0.135.0
jinja2>=3.1.0
uvicorn[standard]>=0.44.0
python-multipart>=0.0.24
pyyaml>=6.0.1
minio>=7.2.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# Database
motor>=3.3.0
pymongo>=4.6.0
beanie>=1.24.0

# AI/ML dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
pillow>=10.0.0
# git+https://github.com/facebookresearch/segment-anything.git
# git+https://github.com/openai/CLIP.git

# 3D processing
# open3d>=0.17.0  # Not compatible with Python 3.13 yet

# Physics
pybullet>=3.2.5

# Task queue
celery>=5.3.0
redis>=5.0.0
